+++ sekoia_automation/scripts/documentation/generate.py
@@ -86,6 +86,12 @@ class DocumentationGenerator:
         # Copy the logo
         module_logo_filename = self._copy_logo(module_path, module_manifest)
 
+        # Copy the assets
+        assets_location = self._copy_assets(module_path, module_manifest)
+
+        # Copy content of CONFIGURE.md
+        setup = self._copy_configure_md(module_path, assets_location)
+
         current_dir = Path(os.path.dirname(__file__)).resolve()
         file_loader = FileSystemLoader(current_dir / "templates")
         env = Environment(loader=file_loader)
@@ -96,6 +102,7 @@ class DocumentationGenerator:
             actions=module_actions,
             triggers=module_triggers,
             logo_filename=module_logo_filename,
+            setup=setup,
         )
 
         module_doc_filename = f'{slugify(module_manifest["name"])}.md'
@@ -131,6 +138,50 @@ class DocumentationGenerator:
                 )
                 return module_logo_filename
         return None
+
+    def _copy_assets(self, module_path: Path, module_manifest: dict) -> str | None:
+        """
+        Copy the assets from the module directory to the documentation directory.
+        Return the new location of the assets
+        """
+        assets_dir = "docs/assets"
+        assets_path = module_path / assets_dir
+
+        if assets_path.exists():
+            new_location = (
+                f"assets/playbooks/library/setup/{slugify(module_manifest['name'])}"
+            )
+            new_location_path = self.documentation_path / "docs" / new_location
+            new_location_path.mkdir(parents=True, exist_ok=True)
+
+            for item in assets_path.iterdir():
+                copyfile(item, new_location_path / item.name)
+
+            return new_location
+
+        return None
+
+    def _copy_configure_md(
+        self, module_path: Path, assets_location: str | None
+    ) -> str | None:
+        """
+        Copy and return the CONFIGURE.md content
+
+        Read the content of CONFIGURE.md in the automation module directory.
+        """
+        # Is the SETUP.md exists
+        if not (module_path / "CONFIGURE.md").exists():
+            return None
+
+        # Read it content
+        with (module_path / "CONFIGURE.md").open("r") as fd:
+            content = fd.read()
+
+        if assets_location:
+            assets_dir = "docs/assets"
+            content = content.replace(assets_dir, f"/{assets_location}")
+
+        return content
 
     def _update_submenu(
         self, menu: list[dict], target: str, value: list, append_only: bool = False