+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add a default logo in the template used to generate new modules
 - Remove `[]` around `Unreleased` entry in the changelog
+- Include CONFIGURE.md into the generated documentation of automation modules
 
 ## 1.19.3 - 2025-08-04